It works ! The +0.5 -0.5 trick appears to work fine !
Still, this is insane. However, it works with GL_LINEAR AND
Multisample. It even gets better with Multisample, which is great.

The +0.5 -0.5 seems to be due to the fact that 'fragment's datum are
sampled from the fragment barycenter', and the fact that GL_LINEAR
will use 2 adjacent pixels when filtering the texture.

This is still not very clear to me, but it's enough that it works
fine.

See this for more details :
https://community.arm.com/message/39319#39319

Signed-off-by: Myy <myy@miouyouyou.fr>