Tex coords in USHORT. Because CPU can't Half-float :C
Texture coordinates in unsigned short offer a good compromise between
precision (1.52520399603e-05), calculation (Every cpu handle halfwords),
and bandwidth (2 bytes against 4 bytes for float). This is the only
format that offers these three items in the CPU !

Most CPU CANNOT handle half-float natively. They can only convert values
to and from half-float. GPU however can do computation with half-floats.
Since I'd like to compute the coordinates of the texture to display,
using the the card value as a base and the texture atlas as a domain,
Unsigned short seems to be the best compromise.

I could use a pre-calculated half-float array but this would take more
memory.

The second problem with half-floats is that they are not available in
OpenGL ES 2.x core. They are available as an extension, but the
availability of this extension is not guaranteed.

If these were coordinates that could change every frame, I might have
chosen to compute these values from the vertex shader, but here it
will be overkill. One time computation and store is slightly better.

Anyway, reduced overhead ! Woohoo ! 144 bytes per card instead of :
- 288 bytes (xy  st)
- 360 bytes (xyz st)
However, I need to check the visuals of every card before being sure
that is appropriate.

Signed-off-by: Myy <myy@miouyouyou.fr>